# konig-urban-backend/services/orders-service/openapi.yaml
openapi: 3.0.0
info:
  title: König Urban - Orders Service
  description: Microservicio encargado de la gestión de pedidos, procesamiento de pagos (Stripe) y logística de envíos.
  version: 1.0.0

servers:
  - url: /api/v1/orders
    description: Orders Service Base URL

tags:
  - name: Shopper Orders
    description: Gestión de pedidos para el cliente final.
  - name: Shipments
    description: Seguimiento y estado de envíos.
  - name: Logistics (ERP)
    description: Gestión de almacén, picking y despacho para administradores.

paths:
  /:
    post:
      tags: [Shopper Orders]
      summary: Crear un nuevo pedido (Checkout)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreateInput" }
      responses:
        "201":
          description: Pedido creado, se retorna la sesión de pago.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  stripeSessionUrl: { type: string, format: uri }
    get:
      tags: [Shopper Orders]
      summary: Listar mis pedidos
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Historial de pedidos del cliente actual.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }

  /{orderId}:
    get:
      tags: [Shopper Orders]
      summary: Detalle de un pedido específico
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del pedido.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }

  /shipments/{orderId}:
    get:
      tags: [Shipments]
      summary: Consultar estado del envío de un pedido
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Información de logística.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }

  /erp/orders:
    get:
      tags: [Logistics (ERP)]
      summary: Listar todos los pedidos para gestión (ERP)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            {
              type: string,
              enum:
                [pendiente, pagado, procesando, enviado, entregado, cancelado],
            }
      responses:
        "200":
          description: Listado global de pedidos.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }

  /erp/orders/{orderId}/status:
    patch:
      tags: [Logistics (ERP)]
      summary: Actualizar estado de pedido o envío (ERP)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                estado: { type: string }
                trackingUrl: { type: string }
                operadorLogistico: { type: string }
      responses:
        "200":
          description: Estado actualizado correctamente.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        clienteId: { type: string, format: uuid }
        fecha: { type: string, format: date-time }
        estado:
          {
            type: string,
            enum:
              [pendiente, pagado, procesando, enviado, entregado, cancelado],
          }
        total: { type: number, format: decimal }
        metodoPago: { type: string }
        createdAt: { type: string, format: date-time }

    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            lineas:
              type: array
              items: { $ref: "#/components/schemas/OrderItem" }
            envio: { $ref: "#/components/schemas/Shipment" }

    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        prendaId: { type: string, format: uuid }
        uidPrendaId: { type: string, format: uuid, nullable: true }
        cantidad: { type: integer }
        precioUnitario: { type: number, format: decimal }

    OrderCreateInput:
      type: object
      required: [lineas, metodoPago]
      properties:
        metodoPago: { type: string }
        lineas:
          type: array
          items:
            type: object
            required: [prendaId, cantidad]
            properties:
              prendaId: { type: string, format: uuid }
              cantidad: { type: integer, minimum: 1 }

    Shipment:
      type: object
      properties:
        id: { type: string, format: uuid }
        pedidoId: { type: string, format: uuid }
        operadorLogistico: { type: string }
        trackingUrl: { type: string }
        estadoEnvio:
          { type: string, enum: [pendiente, en_transito, entregado, devuelto] }
        fechaEnvio: { type: string, format: date-time, nullable: true }
        fechaEntregaEstimada: { type: string, format: date, nullable: true }
        packagingPremium: { type: boolean }